Always enable buttons of remote forms after ajax response.
If a remote form was submitted, the returning function had to handle the re-enabling of the submit button.
This was not visible because at normal cases the content of the form was replaced from the ajax response.
But sometimes this is not happening and the button had to be re-enabled manually.

Bonus: This also includes a fix for buttons with icons, that loses it's icon after getting disabled.